wave 2 methods implimented
