+++ spec/hotel_spec.rb
@@ -1,0 +1,11 @@
+require_relative 'spec_helper'
+
+describe "hotel" do
+  # it 'initialize' do
+  #
+  # end
+
+  it 'get reservations inclusive' do
+
+  end
+end
