+++ spec/spec_helper.rb
@@ -6,3 +6,5 @@ require 'minitest/reporters'
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 # Require_relative your lib files here!
+require_relative '../lib/hotel'
+require_relative '../lib/reservation'
